Framework: split step definitions using Domain concept
