+++ CucumberBDD-OC/src/test/java/framework/context/TestContext.java
@@ -1,7 +1,9 @@
 package framework.context;
 
+import framework.domainObjects.BillingDetails;
 import org.openqa.selenium.WebDriver;
 
 public class TestContext {
     public WebDriver driver;
+    public BillingDetails billingDetails;
 }
